Address addtional CR comments: cleanups, adjusting some APIs between base and derived classes; Remaining CR items are related to ongoing refactoring discussion
