+++ src/Microsoft.OData.Core/Build.NetStandard/Microsoft.OData.Core.NetStandard.csproj
@@ -478,7 +478,10 @@
     </Compile>
     <Compile Include="..\MultipartMixed\ODataMultipartMixedBatchWriter.cs">
       <Link>MultipartMixed\ODataMultipartMixedBatchWriter.cs</Link>
-    </Compile>      
+    </Compile>
+    <Compile Include="..\MultipartMixed\ODataMultipartMixedBatchWriterUtils.cs">
+      <Link>MultipartMixed\ODataMultipartMixedBatchWriterUtils.cs</Link>
+    </Compile>    
     <Compile Include="..\NonDisposingStream.cs">
       <Link>NonDisposingStream.cs</Link>
     </Compile>
@@ -547,9 +550,6 @@
     </Compile>
     <Compile Include="..\ODataBatchWriter.cs">
       <Link>ODataBatchWriter.cs</Link>
-    </Compile>
-    <Compile Include="..\ODataBatchWriterUtils.cs">
-      <Link>ODataBatchWriterUtils.cs</Link>
     </Compile>
     <Compile Include="..\ODataCollectionReader.cs">
       <Link>ODataCollectionReader.cs</Link>
