+++ test/FunctionalTests/Tests/DataOData/Tests/OData.Common.Tests/PublicApi/PublicApi.bsl
@@ -4087,25 +4087,22 @@ public abstract class Microsoft.OData.ODataBatchWriter : IODataBatchOperationLis
 	public Microsoft.OData.ODataBatchOperationResponseMessage CreateOperationResponseMessage (string contentId)
 	public System.Threading.Tasks.Task`1[[Microsoft.OData.ODataBatchOperationResponseMessage]] CreateOperationResponseMessageAsync (string contentId)
 	protected abstract Microsoft.OData.ODataBatchOperationResponseMessage CreateOperationResponseMessageImplementation (string contentId)
-	protected abstract void DisposeBatchWriterAndSetContentStreamRequestedState ()
 	public abstract void Flush ()
 	public abstract System.Threading.Tasks.Task FlushAsync ()
 	public abstract void OnInStreamError ()
 	protected abstract void SetState (Microsoft.OData.ODataBatchWriter+BatchWriterState newState)
 	protected abstract void StartBatchOperationContent ()
-	protected abstract void ValidateWriterReady ()
 	protected abstract void VerifyCanCreateOperationRequestMessage (bool synchronousCall, string method, System.Uri uri, string contentId)
 	protected abstract void VerifyCanFlush (bool synchronousCall)
+	protected abstract void VerifyNotDisposed ()
 	public void WriteEndBatch ()
 	public System.Threading.Tasks.Task WriteEndBatchAsync ()
 	protected abstract void WriteEndBatchImplementation ()
 	public void WriteEndChangeset ()
 	public System.Threading.Tasks.Task WriteEndChangesetAsync ()
 	protected abstract void WriteEndChangesetImplementation ()
-	protected abstract void WritePendingMessageData (bool reportMessageCompleted)
 	public void WriteStartBatch ()
 	public System.Threading.Tasks.Task WriteStartBatchAsync ()
-	protected abstract void WriteStartBoundaryForOperation ()
 	public void WriteStartChangeset ()
 	public System.Threading.Tasks.Task WriteStartChangesetAsync ()
 	protected abstract void WriteStartChangesetImplementation ()
